auka-system/requirements.txt
# AUKA System v2.0 - Dependencias
# Instalar: pip install -r requirements.txt

# ── Core ──────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
python-json-logger>=2.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0

# ── Database ──────────────────────────────────────────────────────
supabase>=2.0.0
aiosqlite>=0.19.0

# ── Telegram Bot ──────────────────────────────────────────────────
python-telegram-bot>=20.0

# ── Scraping ──────────────────────────────────────────────────────
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# ── LLM / API Clients ─────────────────────────────────────────────
openai>=1.0.0  # Compatible con múltiples proveedores (Kimi, DeepSeek, Gemini)
google-generativeai>=0.5.0  # Google Gemini API

# ── Testing ───────────────────────────────────────────────────────
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
